misc/HtmlTestSuite/results/test-0993.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml"><head><title>Historical embed element features should not be supported</title>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
</head><body><div id="log"/>
<embed id="embed"/>
<script>
test(function() {
  var elm = document.getElementById('embed');
  assert_equals(typeof elm, 'object', 'typeof');
  assert_throws_js(TypeError, function() {
    elm();
  });
}, 'embed legacycaller should not be supported');
</script>
</body></html>
